tasks-manager-v2/requirements.txt
Flask
redis
load_dotenv
prometheus-flask-exporter